requirements.txt
Django>=6.0.2
gunicorn>=25.1.0
dj-database-url>=3.1.2
psycopg2-binary>=2.9.11
python-decouple>=3.8
django-crontab>=0.7.1
django-q2>=1.9.0
mailersend==2.0.0
django-ratelimit>=4.1.0
tenacity>=9.1.4
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-simple-captcha==0.6.3
user-agents==2.2.0
Pillow==12.1.1
supabase==2.28.0
whitenoise==6.12.0